+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - Fixed a bug where the `system.schemaVersion` was not getting set when rebuilding the project config.
 - Fixed a bug where all file caches were being saved inside a single subdirectory of `storage/runtime/cache/` based on the cache key prefix. ([#6746](https://github.com/craftcms/cms/issues/6746))
 - Fixed a bug where field handles would get jumpy on hover if their increased size caused them to be too large to fit on the same line as the field name. ([#6742](https://github.com/craftcms/cms/issues/6742))
+- Fixed a bug where the “Photo” field was shown on Edit User pages, even if the User Photos volume had been deleted. ([#6752](https://github.com/craftcms/cms/issues/6752))
 
 ## 3.5.7 - 2020-08-26